Update and rename StringLength.java to length-of-string.java
